% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{Anno2Table}
\alias{Anno2Table}
\title{Convert annotation object to pathway table}
\usage{
Anno2Table(AnnoEnv)
}
\arguments{
\item{AnnoEnv}{An \code{EnrichAnno} object or legacy annotation environment.}
}
\value{
Data frame with pathway metadata and concatenated genes.
}
\description{
Convert annotation object to pathway table
}
