% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_core.R
\name{anno_env_to_EnrichAnno}
\alias{anno_env_to_EnrichAnno}
\title{Internal helper to convert a legacy annotation environment to EnrichAnno.}
\usage{
anno_env_to_EnrichAnno(anno_env)
}
\arguments{
\item{anno_env}{Environment with PATHID2EXTID/PATHID2NAME.}
}
\value{
\code{EnrichAnno}.
}
\description{
Internal helper to convert a legacy annotation environment to EnrichAnno.
}
