% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_core.R
\name{anno_term_description}
\alias{anno_term_description}
\title{Internal helper to map term id to description.}
\usage{
anno_term_description(term, annoData)
}
\arguments{
\item{term}{Character vector of term IDs.}

\item{annoData}{\code{EnrichAnno}.}
}
\value{
Character vector.
}
\description{
Internal helper to map term id to description.
}
