% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_core.R
\name{anno_to_geneSetList}
\alias{anno_to_geneSetList}
\title{Internal helper to build a named gene-set list from EnrichAnno.}
\usage{
anno_to_geneSetList(annoData)
}
\arguments{
\item{annoData}{\code{EnrichAnno}.}
}
\value{
Named list of gene vectors.
}
\description{
Internal helper to build a named gene-set list from EnrichAnno.
}
