% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ask_user_continue}
\alias{ask_user_continue}
\title{Prompt User to Continue with Yes/No Input}
\usage{
ask_user_continue(msg)
}
\arguments{
\item{msg}{A character string containing the message to display to the user before the prompt.}
}
\value{
A logical value: \code{TRUE} if the user inputs 'Y' (Yes), and \code{FALSE} if the user inputs 'N' (No).
}
\description{
This function continuously prompts the user with a message asking whether they want to continue.
It accepts either 'Y' (for Yes) or 'N' (for No) as valid inputs. The function will keep prompting
the user until they provide valid input.
}
