% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_data}
\alias{brain_data}
\title{Effect sizes map for MDD (Left Hemisphere)}
\format{
A data frame with rows as regions and columns as effect sizes of case-control comparisons on regional cortical thickness between major depressive disorders and healthy controls.
}
\source{
}
\usage{
brain_data
}
\description{
A data frame with rows as regions and columns as effect sizes of case-control comparisons on regional cortical thickness between major depressive disorders and healthy controls.
}
\keyword{datasets}
