% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainscore.hpc.R
\name{combine.rds}
\alias{combine.rds}
\title{Combine Results from Saved RDS Files}
\usage{
combine.rds(
  output_dir,
  n_rds = NULL,
  save_name = NULL,
  delete_originals = TRUE
)
}
\arguments{
\item{output_dir}{A character string specifying the directory where the RDS files are stored.}

\item{n_rds}{An integer specifying the expected number of RDS files. If provided, the function will check if any files are missing.}

\item{save_name}{A character string specifying the name of the output RDS file for the combined results. Default is "combined_results.rds".}

\item{delete_originals}{A logical indicating whether to delete the original RDS files after combining. Default is TRUE.}
}
\value{
The combined results, either as a saved RDS file or returned directly if \code{save_combined} is FALSE.
}
\description{
This function combines results from multiple saved RDS files into a single data frame or list,
saves the combined results as an RDS file (optional), and checks for missing files.
}
