% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{filter_geneSetList}
\alias{filter_geneSetList}
\title{Filter gene set list}
\usage{
filter_geneSetList(bg_genes, geneSetList, minGSSize, maxGSSize)
}
\arguments{
\item{bg_genes}{Character vector of background genes.}

\item{geneSetList}{Named list of gene sets.}

\item{minGSSize}{Minimum size threshold.}

\item{maxGSSize}{Maximum size threshold.}
}
\value{
Filtered named list of gene sets.
}
\description{
This function filters a list of gene sets based on background genes and size.
}
