% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{FilterTable}
\alias{FilterTable}
\title{Filter and intersect genes in annotation table}
\usage{
FilterTable(df, bg_genes, sep = ";", minGSSize = 0, maxGSSize = Inf)
}
\arguments{
\item{df}{Data frame with \code{pathID}, \code{pathName}, \code{geneID}.}

\item{bg_genes}{Character vector of background genes.}

\item{sep}{Gene separator in \code{geneID}.}

\item{minGSSize}{Minimum retained gene set size.}

\item{maxGSSize}{Maximum retained gene set size.}
}
\value{
Filtered annotation table.
}
\description{
Filter and intersect genes in annotation table
}
