% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_annoData}
\alias{get_annoData}
\title{Load gene set annotation data}
\usage{
get_annoData(
  type = c("CellTypes_Lake2018", "CellTypes_Martins2021", "CellTypes_Seidlitz2020",
    "DGN", "GO_BP", "GO_CC", "GO_MF", "KEGG", "MKEGG", "Reactome", "SynGO",
    "WikiPathways", "MeSH_A", "MeSH_B", "MeSH_C", "MeSH_D", "MeSH_E", "MeSH_F", "MeSH_G",
    "MeSH_H", "MeSH_I", "MeSH_J", "MeSH_K", "MeSH_L", "MeSH_M", "MeSH_N", "MeSH_Z")
)
}
\arguments{
\item{type}{A character string specifying the type of gene set to load.}
}
\value{
An object of class \code{EnrichAnno}.
}
\description{
This function loads annotation data from RDS files in package data resources.
If data is not found in packaged resources, it attempts to download the file
to the user cache directory from the BrainEnrichData companion package.
}
\examples{
\dontrun{
annoData <- get_annoData("GO_BP")
}
}
