% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneExp.R
\name{get_geneExp}
\alias{get_geneExp}
\title{Get gene expression data}
\usage{
get_geneExp(
  atlas = c("desikan", "schaefer100", "schaefer200", "schaefer300"),
  rdonor = c("r0.2", "r0.4", "r0.6"),
  hem = c("L", "R", "B")
)
}
\arguments{
\item{atlas}{Atlas name. One of "desikan", "schaefer100", "schaefer200", "schaefer300".}

\item{rdonor}{Donor threshold. One of "r0.2", "r0.4", "r0.6".}

\item{hem}{Hemisphere selection. One of "L", "R", "B".}
}
\value{
A matrix containing the gene expression data.
}
\description{
This function retrieves expression matrices for supported atlases. It first
looks for packaged resources (preferring \code{BrainEnrichData}), including an
installed companion package or a sibling development checkout, and falls
back to downloading into the user cache directory.
}
\examples{
\dontrun{
geneExpMatrix <- get_geneExp("desikan", "r0.4", "L")
}
}
