% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_geneSetList}
\alias{get_geneSetList}
\title{Get gene set list}
\usage{
get_geneSetList(annoData)
}
\arguments{
\item{annoData}{Annotation object of class \code{EnrichAnno}.}
}
\value{
A named list of gene sets.
}
\description{
Get gene set list
}
