% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_termDescription}
\alias{get_termDescription}
\title{Get gene set descriptions}
\usage{
get_termDescription(term, annoData, strip_prefix = "")
}
\arguments{
\item{term}{Term IDs.}

\item{annoData}{An object of class \code{EnrichAnno}.}

\item{strip_prefix}{A character string to remove from the beginning of each term.}
}
\value{
A character vector of gene set descriptions.
}
\description{
Get gene set descriptions
}
