% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_core.R
\name{is_EnrichAnno}
\alias{is_EnrichAnno}
\title{Check if object is EnrichAnno}
\usage{
is_EnrichAnno(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical scalar.
}
\description{
Check if object is EnrichAnno
}
