% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_core.R
\name{new_EnrichAnno}
\alias{new_EnrichAnno}
\title{Construct an EnrichAnno object}
\usage{
new_EnrichAnno(term2gene, term2name, meta = list())
}
\arguments{
\item{term2gene}{Data frame with columns \code{term_id} and \code{gene_id}.}

\item{term2name}{Data frame with columns \code{term_id} and \code{term_name}.}

\item{meta}{Optional named list of metadata.}
}
\value{
An object of class \code{EnrichAnno}.
}
\description{
Construct an EnrichAnno object
}
