% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perm_id_dk_lh_5000}
\alias{perm_id_dk_lh_5000}
\title{Permutation index for left Desikan regions (5000 permutations)}
\format{
A matrix with rows as regions and columns as permutated indices.
}
\source{
rotate_parcellation(coord.l = coord_dk_lh, nrot = 5000)
}
\usage{
perm_id_dk_lh_5000
}
\description{
Permutation index for left Desikan regions (5000 permutations)
}
\keyword{datasets}
