% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_brain}
\alias{plot_brain}
\title{Plot Brain Data}
\usage{
plot_brain(
  df2plot,
  ats = c("dx", "dk", "aseg"),
  what2plot = "statistic",
  filterby = c("p.value", "p.adj", "none"),
  title2show = "",
  limit2show = c(-15, 15),
  legend2show = "Stat",
  hide_legend = FALSE,
  hem = "both",
  low = "steelblue1",
  mid = "white",
  high = "firebrick1",
  sufix2remove = "_thickavg"
)
}
\arguments{
\item{df2plot}{A data frame containing the data to plot.}

\item{ats}{A character string indicating the atlas to use ('dx', 'dk', 'aseg').}

\item{what2plot}{A character string indicating the variable to plot ('statistic').}

\item{filterby}{A character string indicating the filter to apply ('p.value', 'p.adj', 'none').}

\item{title2show}{A character string indicating the title of the plot.}

\item{limit2show}{A numeric vector of length 2 indicating the limits for the color scale.}

\item{legend2show}{A character string indicating the legend title.}

\item{hide_legend}{A logical value indicating whether to hide the legend.}

\item{hem}{A character string indicating which hemisphere to plot ('both', 'left', 'right').}

\item{low}{A character string indicating the color for the low end of the scale.}

\item{mid}{A character string indicating the color for the midpoint of the scale.}

\item{high}{A character string indicating the color for the high end of the scale.}

\item{sufix2remove}{A character string indicating the suffix to remove from labels.}
}
\value{
A ggplot2 object.
}
\description{
This function creates a brain plot using the ggplot2 and ggseg packages.
}
