% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_core_genes}
\alias{plot_core_genes}
\title{Plot core genes for a term}
\usage{
plot_core_genes(x, term_id, mode = c("impact", "direction"))
}
\arguments{
\item{x}{An \code{EnrichRes} object.}

\item{term_id}{Term identifier.}

\item{mode}{Plot either core-gene \code{"impact"} or role \code{"direction"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot core genes for a term
}
