% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap_terms}
\alias{plot_heatmap_terms}
\title{Plot a permutation heatmap for top terms}
\usage{
plot_heatmap_terms(x, top_n = 20, max_perm = 100)
}
\arguments{
\item{x}{An \code{EnrichRes} object.}

\item{top_n}{Number of terms to display.}

\item{max_perm}{Maximum number of permutations to display.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a permutation heatmap for top terms
}
