% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainscore.R
\name{simple_lm}
\alias{simple_lm}
\title{Perform Linear Regression with Multiple Predictors and Covariates}
\usage{
simple_lm(
  dependent_df,
  pred_df,
  cov_df,
  stat2return = c("full", "statistic", "p.value")
)
}
\arguments{
\item{stat2return}{A character string specifying which statistic to return ("statistic", "p.value", or "full"). Default is "full". "statistic" returns only the t-value for permutation purpose, "p.value" returns only the p-value for simulation analysis, and "full" returns all information for the parametric test.}

\item{df}{A data frame containing the data.}

\item{dependent_vars}{A character vector of dependent variable names.}

\item{pred_var}{A character vector of predictor variable names.}

\item{cov_vars}{A character vector of covariate variable names.}

\item{var2extract}{A character string or vector specifying which predictor's coefficient to extract. If NULL, defaults to pred_var. For a factor variable, specify the level to extract.}
}
\value{
A data frame containing model summaries.
}
\description{
This function fits linear models for specified dependent variables using given predictors and covariates.
It returns a data frame containing model summaries.
}
