% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{split_Anno}
\alias{split_Anno}
\title{Split annotation data}
\usage{
split_Anno(AnnoEnv)
}
\arguments{
\item{AnnoEnv}{An \code{EnrichAnno} object or legacy annotation environment.}
}
\value{
A list with \code{path2gene} and \code{path2name} data frames.
}
\description{
Convert an annotation object to path2gene/path2name data frames.
}
