% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{Table2Anno}
\alias{Table2Anno}
\title{Convert a data frame to EnrichAnno}
\usage{
Table2Anno(df, sep = ";")
}
\arguments{
\item{df}{A data frame with columns \code{pathID}, \code{pathName}, and \code{geneID}.}

\item{sep}{Gene ID separator in \code{geneID}.}
}
\value{
An object of class \code{EnrichAnno}.
}
\description{
Convert a data frame to EnrichAnno
}
