% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_data}
\alias{brain_data}
\title{Brain Data PC1 for Left Hemisphere}
\format{
A data frame with rows as regions and columns as effect sizes of case-control comparisons on regional cortical thickness between bipolar disorders and healthy controls.
}
\source{
read.csv('data-raw/desikan_PC1_data.csv')
}
\usage{
brain_data
}
\description{
This dataset contains PC1 data filtered for regions starting with 'L_' in the Desikan atlas.
}
\keyword{datasets}
