% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainscore.R
\name{brainscore.simulate}
\alias{brainscore.simulate}
\title{Perform Brain Score Simulation}
\usage{
brainscore.simulate(
  pred_df,
  cov_df,
  brain_data,
  gene_data,
  annoData,
  sim_n = 1000,
  subsample_size = 100,
  sim_type = c("randomize_pred", "spin_brain", "resample_gene", "coexp_matched"),
  cor_method = c("pearson", "spearman", "pls1c", "pls1w", "custom"),
  aggre_method = c("mean", "median", "meanabs", "meansqr", "maxmean", "ks_orig",
    "ks_weighted", "ks_pos_neg_sum", "sign_test", "rank_sum", "custom"),
  null_model = c("spin_brain", "resample_gene"),
  minGSSize = 10,
  maxGSSize = 200,
  n_cores = 0,
  n_perm = 5000,
  perm_id = NULL
)
}
\arguments{
\item{pred_df}{Data frame of predictor variables.}

\item{cov_df}{Data frame of covariate variables.}

\item{brain_data}{Data frame of brain imaging data.}

\item{gene_data}{Data frame of gene expression data.}

\item{annoData}{Environment containing annotation data.}

\item{sim_n}{Integer specifying the number of simulations. Default is 1000.}

\item{subsample_size}{Integer or vector specifying the subsample sizes. Default is 100.}

\item{sim_type}{Character string specifying the simulation type. Default is 'randomize_pred'.
Other options include 'spin_brain', 'resample_gene', 'coexp_matched'.}

\item{cor_method}{Character string specifying the correlation method. Default is 'pearson'.
Other options include 'spearman', 'pls1c', 'pls1w', 'custom'.}

\item{aggre_method}{Character string specifying the aggregation method. Default is 'mean'.
Other options include 'median', 'meanabs', 'meansqr', 'maxmean',
'ks_orig', 'ks_weighted', 'ks_pos_neg_sum', 'sign_test', 'rank_sum', 'custom'.}

\item{null_model}{Character string specifying the null model method. Default is 'spin_brain'.
Other option is 'resample_gene'.}

\item{minGSSize}{Integer specifying the minimum gene set size. Default is 10.}

\item{maxGSSize}{Integer specifying the maximum gene set size. Default is 200.}

\item{n_cores}{Integer specifying the number of cores to use for parallel processing. Default is 0.}

\item{n_perm}{Integer specifying the number of permutations. Default is 5000.}

\item{perm_id}{Optional permutation ID.}
}
\value{
A list of data frames containing the results of the simulations.
}
\description{
This function performs simulations on brain score data using different methods for comparison.
It calculates gene set scores, performs linear modeling, and returns the simulation results.
}
