% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{FilterTable}
\alias{FilterTable}
\title{Filter and Intersect Genes in Annotation Data}
\usage{
FilterTable(df, bg_genes, sep = ";", minGSSize = 0, maxGSSize = Inf)
}
\arguments{
\item{df}{A dataframe containing the annotation information with columns:
\itemize{
\item \code{pathID}: The ID of the pathway/term.
\item \code{pathName}: The descriptive name of the pathway/term.
\item \code{geneID}: A string of gene identifiers associated with the pathway/term, separated by a delimiter.
}}

\item{sep}{A character string that defines the delimiter for splitting the \code{geneID} column. Defaults to \code{";"}.}

\item{minGSSize}{An integer specifying the minimum gene set size to retain. Defaults to \code{0}.}

\item{maxGSSize}{An integer specifying the maximum gene set size to retain. Defaults to \code{Inf}.}

\item{background}{A character vector of background gene names.}
}
\value{
A dataframe with columns:
\itemize{
\item \code{pathID}: The ID of the pathway/term.
\item \code{pathName}: The descriptive name of the pathway/term.
\item \code{geneSetSize}: The size of the filtered gene set.
\item \code{geneID}: The filtered and concatenated gene identifiers.
}
}
\description{
This function filters the genes in the \code{geneID} column of an annotation dataframe
based on a given list of background genes. Only genes present in the background list
are retained in the resulting dataframe. It also calculates and includes the \code{geneSetSize} column,
and filters gene sets based on specified size constraints.
}
