% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_annoData}
\alias{get_annoData}
\title{Load gene sets annotation data}
\usage{
get_annoData(
  type = c("CellTypes_Lake2018", "CellTypes_Martins2021", "CellTypes_Seidlitz2020",
    "DGN", "GO_BP", "GO_CC", "GO_MF", "KEGG", "Reactome", "SynGO", "WikiPathways")
)
}
\arguments{
\item{type}{A character string specifying the type of gene set to load. Options are:
\describe{
\item{"CellTypes_Lake2018"}{Cell types data from Lake2018}
\item{"CellTypes_Martins2021"}{Cell types data from Martins2021}
\item{"CellTypes_Seidlitz2020"}{Cell types data from Seidlitz2020}
\item{"DGN"}{DisGeNET gene sets}
\item{"GO_BP"}{Gene Ontology Biological Process}
\item{"GO_CC"}{Gene Ontology Cellular Component}
\item{"GO_MF"}{Gene Ontology Molecular Function}
\item{"KEGG"}{KEGG gene sets}
\item{"Reactome"}{Reactome gene sets}
\item{"SynGO"}{SynGO gene sets}
\item{"WikiPathways"}{WikiPathways gene sets}
}}
}
\value{
A data frame containing the annotation data.
}
\description{
This function loads annotation data from RDS files located in the \code{inst/extdata/geneSets} directory of the package.
If the specified file does not exist locally, it will be downloaded from the GitHub repository.
}
\examples{
\dontrun{
annoData <- get_annoData("GO_BP")
}
}
