% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_termDescription}
\alias{get_termDescription}
\title{Get Gene Set Descriptions}
\usage{
get_termDescription(term, annoData, strip_prefix = "")
}
\arguments{
\item{term}{term to search from annoData (can be a vector of terms).}

\item{annoData}{An environment containing annotation data.}

\item{strip_prefix}{A character string to remove from the beginning of each term.}
}
\value{
A character vector of gene set descriptions.
}
\description{
This function retrieves descriptions for gene sets from annotation data.
}
