% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_core_genes.R
\name{identify_core_genes}
\alias{identify_core_genes}
\title{Identify core genes based on a specified threshold method}
\usage{
identify_core_genes(
  changes,
  threshold_type = c("sd", "percentile"),
  threshold_value = 1
)
}
\arguments{
\item{changes}{Named vector of changes from LOO analysis.}

\item{threshold_value}{Numeric value indicating the percentile (if method is "percentile") or the number of standard deviations (if method is "sd").}

\item{method}{Character string specifying the method to use for threshold calculation ("percentile" or "sd").}
}
\value{
Vector of core genes or NA if no core genes are identified.
}
\description{
Identify core genes based on a specified threshold method
}
