% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{split_Anno}
\alias{split_Anno}
\title{Split Annotation Data}
\usage{
split_Anno(Anno_clusterProfiler_Env)
}
\arguments{
\item{Anno_clusterProfiler_Env}{An environment containing annotation data, including
\code{PATHID2EXTID} and \code{PATHID2NAME}, created by the \code{build_Anno} function.}
}
\value{
A list with two components:
\describe{
\item{path2gene}{A data frame with two columns: \code{pathID} and \code{geneID}, representing
the relationship between pathway IDs and gene IDs.}
\item{path2name}{A data frame with two columns: \code{pathID} and \code{pathName}, representing
the mapping between pathway IDs and their corresponding names. If \code{PATHID2NAME} is not
present in the environment, this component will be \code{NULL}.}
}
}
\description{
This function extracts and reconstructs the \code{path2gene} and \code{path2name} data frames
from an annotation environment created by the \code{build_Anno} function. It retrieves
and organizes pathway IDs and associated gene IDs, as well as pathway names, into
two separate data frames.
}
