% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{Table2Anno}
\alias{Table2Anno}
\title{Convert a Dataframe to Annotation Environment
This function converts a dataframe containing annotation information back to an annotation environment
using the \code{build_Anno} function from the \code{DOSE} package. The resulting environment can be used for
enrichment analysis or other annotation-related tasks.}
\usage{
Table2Anno(df, sep = ";")
}
\arguments{
\item{df}{A dataframe containing the annotation information with columns:
\itemize{
\item \code{pathID}: The ID of the pathway/term.
\item \code{pathName}: The descriptive name of the pathway/term.
\item \code{geneID}: A string of gene identifiers associated with the pathway/term, separated by a delimiter.
}}

\item{sep}{A character string that defines the delimiter for splitting the \code{geneID} column. Defaults to \code{";"}.}
}
\value{
An annotation environment containing the term-to-gene mapping and term descriptions.
}
\description{
Convert a Dataframe to Annotation Environment
This function converts a dataframe containing annotation information back to an annotation environment
using the \code{build_Anno} function from the \code{DOSE} package. The resulting environment can be used for
enrichment analysis or other annotation-related tasks.
}
